backend/pyproject.toml
[project]
name = "youtube-analyzer"
version = "0.1.0"
description = "YouTube video analyzer with AI-powered translation and PPT generation"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "openai-whisper>=20250625",
    "pydantic-settings>=2.13.1",
    "python-pptx>=1.0.2",
    "uvicorn>=0.42.0",
    "websockets>=16.0",
    "yt-dlp>=2026.3.17",
    "torch",
    "torchvision",
    "torchaudio",
    "openai>=2.29.0",
    "comtypes>=1.4.16",
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu128" }
torchvision = { index = "pytorch-cu128" }
torchaudio = { index = "pytorch-cu128" }

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
